# specs/001-esat-skills-portfolio/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Livret Numérique ESAT API
  description: API REST pour l'application de parcours de compétences ESAT
  version: 1.0.0
  contact:
    name: ESAT Support
servers:
  - url: http://localhost:3001/api
    description: Développement local
  - url: https://api.esat-livret.local/api
    description: Production
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Connexion utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=abc123; HttpOnly; Secure; SameSite=Strict
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Déconnexion
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Déconnexion réussie
  /auth/me:
    get:
      tags: [Auth]
      summary: Utilisateur courant
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ USERS (Admin) ============
...
  /events:
    get:
      tags: [Events]
      summary: Mes événements
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Liste des événements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

    post:
      tags: [Events]
      summary: Ajouter un événement
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Événement créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
  /events/{eventId}:
    get:
      tags: [Events]
      summary: Détail d'un événement
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      responses:
        '200':
          description: Détail événement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'

    patch:
      tags: [Events]
      summary: Modifier un événement
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Événement modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'

    delete:
      tags: [Events]
      summary: Supprimer un événement
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      responses:
        '204':
          description: Événement supprimé

  # ============ SKILLS (Référentiel) ============
  /domains:
    get:
      tags: [Skills]
      summary: Liste des domaines avec catégories et compétences
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Hiérarchie complète
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DomainWithHierarchy'

    post:
      tags: [Skills]
      summary: Créer un domaine (Admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainRequest'
      responses:
        '201':
          description: Domaine créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
  /domains/{domainId}:
    patch:
      tags: [Skills]
      summary: Modifier un domaine (Admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/domainId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDomainRequest'
      responses:
        '200':
          description: Domaine modifié

    delete:
      tags: [Skills]
      summary: Supprimer un domaine (Admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/domainId'
      responses:
        '204':
          description: Domaine supprimé
  /categories:
    post:
      tags: [Skills]
      summary: Créer une catégorie (Admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Catégorie créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{categoryId}:
    patch:
      tags: [Skills]
      summary: Modifier une catégorie (Admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Catégorie modifiée

    delete:
      tags: [Skills]
      summary: Supprimer une catégorie (Admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryId'
      responses:
        '204':
          description: Catégorie supprimée
  /skills:
    post:
      tags: [Skills]
      summary: Créer une compétence (Admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSkillRequest'
      responses:
        '201':
          description: Compétence créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
  /skills/{skillId}:
    patch:
      tags: [Skills]
      summary: Modifier une compétence (Admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/skillId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSkillRequest'
      responses:
        '200':
          description: Compétence modifiée

    delete:
      tags: [Skills]
      summary: Supprimer une compétence (Admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/skillId'
      responses:
        '204':
          description: Compétence supprimée

  # ============ USER SKILLS ============
  /my-skills:
    get:
      tags: [UserSkills]
      summary: Mes compétences avec récapitulatif
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Récapitulatif des compétences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MySkillsSummary'
  /my-skills/{skillId}:
    post:
      tags: [UserSkills]
      summary: Ajouter une compétence à mon profil
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/skillId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserSkillRequest'
      responses:
        '201':
          description: Compétence ajoutée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSkill'
        '409':
          description: Compétence déjà ajoutée

    patch:
      tags: [UserSkills]
      summary: Modifier le statut d'une compétence
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/skillId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserSkillRequest'
      responses:
        '200':
          description: Statut modifié

    delete:
      tags: [UserSkills]
      summary: Retirer une compétence de mon profil
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/skillId'
      responses:
        '204':
          description: Compétence retirée

  # ============ VALIDATIONS (Admin) ============
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
...
    eventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    domainId:
      name: domainId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    categoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    skillId:
      name: skillId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Role:
      type: string
      enum: [USER, ADMIN]
    SkillStatus:
      type: string
      enum: [IN_PROGRESS, PENDING_VALIDATION, ACQUIRED, REJECTED]
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        avatarLevel:
          type: integer
          minimum: 1
          maximum: 5
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        title:
          type: string
          maxLength: 100
        photoUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateEventRequest:
      type: object
      required: [date, title]
      properties:
        date:
          type: string
          format: date
        title:
          type: string
          maxLength: 100
        photo:
          type: string
          format: binary
    UpdateEventRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        title:
          type: string
          maxLength: 100
        photo:
          type: string
          format: binary
    Domain:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        displayOrder:
          type: integer
    DomainWithHierarchy:
      allOf:
        - $ref: '#/components/schemas/Domain'
        - type: object
          properties:
            categories:
              type: array
              items:
                $ref: '#/components/schemas/CategoryWithSkills'
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        domainId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        displayOrder:
          type: integer
    CategoryWithSkills:
      allOf:
        - $ref: '#/components/schemas/Category'
        - type: object
          properties:
            skills:
              type: array
              items:
                $ref: '#/components/schemas/Skill'
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        iconName:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
    CreateDomainRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        displayOrder:
          type: integer
    UpdateDomainRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        displayOrder:
          type: integer
    CreateCategoryRequest:
      type: object
      required: [domainId, name]
      properties:
        domainId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        displayOrder:
          type: integer
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        displayOrder:
          type: integer
    CreateSkillRequest:
      type: object
      required: [categoryId, name, description, iconName]
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        iconName:
          type: string
        displayOrder:
          type: integer
    UpdateSkillRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        iconName:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
    UserSkill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        skillId:
          type: string
          format: uuid
        skill:
          $ref: '#/components/schemas/Skill'
        status:
          $ref: '#/components/schemas/SkillStatus'
        requestedAt:
          type: string
          format: date-time
        validatedAt:
          type: string
          format: date-time
          nullable: true
        rejectionReason:
          type: string
          nullable: true
    MySkillsSummary:
      type: object
      properties:
        avatarLevel:
          type: integer
        summary:
          type: object
          properties:
            acquired:
              type: integer
            inProgress:
              type: integer
            pendingValidation:
              type: integer
            rejected:
              type: integer
        skills:
          type: array
          items:
            $ref: '#/components/schemas/UserSkill'
    AddUserSkillRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [IN_PROGRESS, PENDING_VALIDATION]
    UpdateUserSkillRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [IN_PROGRESS, PENDING_VALIDATION]
...